+++ packages/components/src/calendar/date-picker.js
@@ -64,7 +64,7 @@ class DatePicker extends Component {
 	}
 
 	render() {
-		const { date, text, dateFormat, error, isInvalidDate } = this.props;
+		const { date, disabled, text, dateFormat, error, isInvalidDate } = this.props;
 
 		return (
 			<Dropdown
@@ -72,6 +72,7 @@ class DatePicker extends Component {
 				focusOnMount={ false }
 				renderToggle={ ( { isOpen, onToggle } ) => (
 					<DateInput
+						disabled={ disabled }
 						value={ text }
 						onChange={ this.onInputChange }
 						dateFormat={ dateFormat }
@@ -112,6 +113,10 @@ DatePicker.propTypes = {
 	 * A moment date object representing the selected date. `null` for no selection.
 	 */
 	date: PropTypes.object,
+	/**
+	 * Whether the input is disabled.
+	 */
+	disabled: PropTypes.bool,
 	/**
 	 * The date in human-readable format. Displayed in the text input.
 	 */